+++ test/Spec.hs
@@ -3,6 +3,7 @@ module Main (main) where
 import qualified AgentSpec
 import qualified AppSpec
 import qualified CliSpec
+import qualified ContextSpec
 import qualified PolicySpec
 import qualified PromptSpec
 import qualified ProviderSpec
@@ -18,6 +19,7 @@ main = hspec $ do
   describe "Agent" AgentSpec.spec
   describe "App" AppSpec.spec
   describe "Cli" CliSpec.spec
+  describe "Context" ContextSpec.spec
   describe "Policy" PolicySpec.spec
   describe "Prompt" PromptSpec.spec
   describe "Provider" ProviderSpec.spec